+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/tests/Azure.Monitor.OpenTelemetry.AspNetCore.Tests/AzureSdkLoggingTests.cs
@@ -27,19 +27,27 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests
 #else
         [Theory]
 #endif
-        [InlineData(LogLevel.Information, "TestInfoEvent: hello")]
-        [InlineData(LogLevel.Warning, "TestWarningEvent: hello")]
-        [InlineData(LogLevel.Debug, null)]
-        public async Task DistroLogForwarderIsAdded(LogLevel eventLevel, string expectedMessage)
+        [InlineData(false, LogLevel.Debug, null)]
+        [InlineData(false, LogLevel.Information, null)]
+        [InlineData(false, LogLevel.Warning, "TestWarningEvent: hello")]
+        [InlineData(true, LogLevel.Information, "TestInfoEvent: hello")]
+        [InlineData(true, LogLevel.Warning, "TestWarningEvent: hello")]
+        [InlineData(true, LogLevel.Debug, "TestVerboseEvent: hello")]
+        public async Task DistroLogForwarderIsAdded(bool addLoggingFilter, LogLevel eventLevel, string expectedMessage)
         {
             var builder = WebApplication.CreateBuilder();
+            using TestEventSource source = new TestEventSource(addLoggingFilter ? "Azure-LoggingFilter" : "Azure-Test");
+
+            if (addLoggingFilter)
+            {
+                builder.Logging.AddFilter(source.Name.Replace('-', '.'), eventLevel);
+            }
             var transport = new MockTransport(_ => new MockResponse(200).SetContent("ok"));
             SetUpOTelAndLogging(builder, transport, LogLevel.Information);
 
             using var app = builder.Build();
             await app.StartAsync();
 
-            using TestEventSource source = new TestEventSource();
             Assert.True(source.IsEnabled());
             source.LogMessage("hello", eventLevel);
             WaitForRequest(transport);
@@ -79,7 +87,7 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests
 
             await app.StartAsync();
 
-            using TestEventSource source = new TestEventSource();
+            using TestEventSource source = new TestEventSource("Azure-Test");
             Assert.True(source.IsEnabled());
             source.LogMessage("hello", eventLevel);
 
@@ -140,6 +148,99 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests
             Assert.False(logAzureFilterCalled);
         }
 
+        [Fact]
+        public async Task DistroLogForwarderAppliesWildCardFilter()
+        {
+            var builder = WebApplication.CreateBuilder();
+            builder.Logging.AddFilter("Azure.*", LogLevel.Warning);
+
+            var transport = new MockTransport(_ => new MockResponse(200).SetContent("ok"));
+            SetUpOTelAndLogging(builder, transport, LogLevel.Information);
+
+            using var app = builder.Build();
+            await app.StartAsync();
+
+            using TestEventSource source = new TestEventSource("Azure-Test");
+            Assert.True(source.IsEnabled());
+            source.LogMessage("hello", LogLevel.Warning);
+            WaitForRequest(transport);
+
+            Assert.Single(transport.Requests);
+            await AssertContentContains(transport.Requests.Single(), "TestWarningEvent: hello", LogLevel.Warning);
+        }
+
+        [Fact]
+        public async Task SettingCustomLoggingFilterResetsDefaultWarningLevel()
+        {
+            var builder = WebApplication.CreateBuilder();
+            // Even when a single custom filter is set, it should reset the default warning level.
+            builder.Logging.AddFilter("Azure.One", LogLevel.Information);
+
+            var transport = new MockTransport(_ => new MockResponse(200).SetContent("ok"));
+            SetUpOTelAndLogging(builder, transport, LogLevel.Information);
+
+            using var app = builder.Build();
+            await app.StartAsync();
+
+            // Azure-One is added as a logging filter, the default warning level is reset.
+            // Informational-level logs from Azure-One sources are collected.
+            using TestEventSource source1 = new TestEventSource("Azure-One");
+            Assert.True(source1.IsEnabled());
+
+            source1.LogMessage("hello one", LogLevel.Information);
+            WaitForRequest(transport);
+            Assert.Single(transport.Requests);
+            await AssertContentContains(transport.Requests.Single(), "TestInfoEvent: hello one", LogLevel.Information);
+            transport.Requests.Clear();
+
+            // Azure-Two is not part of the logging filter.
+            // Since the logging filter is customized for the Azure SDK, the default warning level is reset.
+            // Informational-level logs from Azure-Two sources are collected.
+            using TestEventSource source2 = new TestEventSource("Azure-Two");
+            Assert.True(source2.IsEnabled());
+
+            source2.LogMessage("hello two", LogLevel.Information);
+            WaitForRequest(transport);
+            Assert.Single(transport.Requests);
+            await AssertContentContains(transport.Requests.Single(), "TestInfoEvent: hello two", LogLevel.Information);
+        }
+
+        [Fact]
+        public async Task CustomLoggingFilterOverridesDefaultWarningAndCapturesErrorLogs()
+        {
+            var builder = WebApplication.CreateBuilder();
+            // Even when a single custom filter is set, it should reset the default warning level.
+            builder.Logging.AddFilter("Azure.One", LogLevel.Error);
+
+            var transport = new MockTransport(_ => new MockResponse(200).SetContent("ok"));
+            SetUpOTelAndLogging(builder, transport, LogLevel.Information);
+
+            using var app = builder.Build();
+            await app.StartAsync();
+
+            using TestEventSource source1 = new TestEventSource("Azure-One");
+            Assert.True(source1.IsEnabled());
+
+            // Only log level with errors should be captured as it is set in the logging filter.
+            source1.LogMessage("Hello Information", LogLevel.Information);
+            source1.LogMessage("Hello Debug", LogLevel.Debug);
+            source1.LogMessage("Hello Warning", LogLevel.Warning);
+            source1.LogMessage("Hello Error", LogLevel.Error);
+            WaitForRequest(transport);
+            Assert.Single(transport.Requests);
+            await AssertContentContains(transport.Requests.Single(), "TestErrorEvent: Hello Error", LogLevel.Error);
+
+            // Azure-Two is not part of the logging filter, it should capture all logs.
+            using TestEventSource source2 = new TestEventSource("Azure-Two");
+            Assert.True(source2.IsEnabled());
+            transport.Requests.Clear();
+
+            source2.LogMessage("hello two", LogLevel.Information);
+            WaitForRequest(transport);
+            Assert.Single(transport.Requests);
+            await AssertContentContains(transport.Requests.Single(), "TestInfoEvent: hello two", LogLevel.Information);
+        }
+
         private IEnumerable<MockRequest> WaitForRequest(MockTransport transport, Func<MockRequest, bool>? filter = null)
         {
             filter = filter ?? (_ => true);
@@ -161,7 +262,8 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests
             contentStream.Position = 0;
             var content = BinaryData.FromStream(contentStream).ToString();
             var jsonMessage = $"\"message\":\"{expectedMessage}\"";
-            var jsonLevel = $"\"severityLevel\":\"{expectedLevel}\"";
+            var level = expectedLevel == LogLevel.Debug ? "Verbose" : expectedLevel.ToString();
+            var jsonLevel = $"\"severityLevel\":\"{level}\"";
             Assert.Contains(jsonMessage, content);
             Assert.Contains(jsonLevel, content);
 
@@ -206,9 +308,11 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests
 
         internal class TestEventSource : AzureEventSource
         {
-            private const string EventSourceName = "Azure-Test";
-            public TestEventSource() : base(EventSourceName)
+            private readonly string EventSourceName;
+
+            public TestEventSource(string eventSourceName) : base(eventSourceName)
             {
+                EventSourceName = eventSourceName;
             }
 
             [Event(1, Level = EventLevel.Informational, Message = "TestInfoEvent: {0}")]
@@ -229,10 +333,19 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore.Tests
                 WriteEvent(3, message);
             }
 
+            [Event(4, Level = EventLevel.Error, Message = "TestErrorEvent: {0}")]
+            public void LogTestErrorEvent(string message)
+            {
+                WriteEvent(4, message);
+            }
+
             public void LogMessage(string message, LogLevel level)
             {
                 switch (level)
                 {
+                    case LogLevel.Error:
+                        LogTestErrorEvent(message);
+                        break;
                     case LogLevel.Warning:
                         LogTestWarningEvent(message);
                         break;